Split out NodalElasticForces
